refactor: implement simple hash-based client side routing
