+++ package.json
@@ -31,5 +31,8 @@
 		"tslib": "^2.6.0",
 		"typescript": "^5.0.2",
 		"vite": "^4.4.5"
+	},
+	"dependencies": {
+		"svelte-spa-router": "^4.0.1"
 	}
 }
