+++ src/lib/Routes/NotFound.svelte
@@ -1,0 +1,1 @@
+<h1>Sorry, this route doesn't exist</h1>
