+++ src/main.ts
@@ -1,4 +1,3 @@
-import "./app.css";
 import App from "./App.svelte";
 
 const app = new App({
